% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keystoneness.R
\name{keystoneness_all}
\alias{keystoneness_all}
\title{Calculate Ecopath-consistent keystoneness indices}
\usage{
keystoneness_all(epu = c("GB", "GOM", "MAB"))
}
\arguments{
\item{epu}{Character. "GB", "GOM", or "MAB"}
}
\value{
data.frame of keystoneness metrics for living groups
}
\description{
Computes Keystone indices #1, #2, #3 and Relative Total Impact
exactly as implemented in Ecopath (Valls et al. + legacy code).
}
