% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MTI.R
\name{MTI}
\alias{MTI}
\title{Calculate Mixed Trophic Impact (MTI)}
\usage{
MTI(Rpath, Rpath.params, increase = TRUE)
}
\arguments{
\item{Rpath}{Rpath object}

\item{Rpath.params}{Rpath parameter object}

\item{increase}{Logical; if TRUE, assumes marginal increase}
}
\value{
MTI matrix
}
\description{
Calculate Mixed Trophic Impact (MTI)
}
