% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mti_new.R
\name{MTI_new}
\alias{MTI_new}
\title{Mixed Trophic Impact (Ecopath-consistent)}
\usage{
MTI_new(Rpath, Rpath.params)
}
\arguments{
\item{Rpath}{A balanced Rpath model object}

\item{Rpath.params}{The corresponding unbalanced/balanced parameter object}
}
\value{
A square MTI matrix with rows and columns equal to all groups
}
\description{
Computes the Mixed Trophic Impact (MTI) matrix for an Rpath model,
reproducing Ecopath behavior including undocumented legacy steps.
}
